Splitting StandardPreconditioners to make it specializable for GPU
