+++ CMakeLists.txt
@@ -756,6 +756,7 @@ if(CUDA_FOUND)
                        gpu_smart_pointers
                        is_gpu_pointer
                        throw_macros_on_gpu
+                       preconditioner_factory_gpu
                        PROPERTIES LABELS ${gpu_label})
   if(MPI_FOUND)
     set_tests_properties(GpuOwnerOverlapCopy PROPERTIES LABELS ${gpu_label})
